DOMN: Put Moves in DOMN namespace
