+++ CTRL/include/CTRL/Controller.hpp
@@ -12,11 +12,11 @@ class Controller
 public:
     Controller(std::shared_ptr<IDriver> driver) : m_driver(driver) { }
 
-    void executeMove(const move::Move& move);
+    void executeMove(const DOMN::Move& move);
 
 private:
-    void executeMove(const move::LinearMove& move);
-    void executeMove(const move::TriangularMove& move);
+    void executeMove(const DOMN::LinearMove& move);
+    void executeMove(const DOMN::TriangularMove& move);
 
     std::shared_ptr<IDriver> m_driver;
 };
